+++ Common/Main_Menu.tscn
@@ -119,6 +119,7 @@ icon_alignment = 1
 language = "en"
 
 [node name="Dialogue" type="Button" parent="Control/MarginContainer/VBoxContainer"]
+unique_name_in_owner = true
 texture_filter = 1
 layout_mode = 2
 size_flags_horizontal = 5
